+++ src/services/S3.php
@@ -3,7 +3,6 @@
 namespace codemonauts\aws\services;
 
 use Aws\S3\S3Client;
-use codemonauts\aws\Aws;
 use codemonauts\aws\traits\S3Trait;
 use yii\base\Component;
 
@@ -14,33 +13,20 @@ class S3 extends Component
     /**
      * @var S3Client The client from the SDK
      */
-    private $client;
+    private S3Client $client;
 
     /**
      * @var string The AWS access key to use, empty to use instance role
      */
-    private $key = '';
+    private string $key = '';
 
     /**
      * @var string The AWS secret
      */
-    private $secret = '';
+    private string $secret = '';
 
     /**
      * @var string The AWS region to use for bucket operations
      */
-    private $region = '';
-
-    /**
-     * @inheritdoc
-     */
-    public function init(): void
-    {
-        $config = Aws::getInstance()->getSettings();
-        $this->key = $config->key;
-        $this->secret = $config->secret;
-        $this->region = $config->region;
-
-        $this->client = $this->getClient();
-    }
+    private string $region = '';
 }